Split the crawler into several modules, add documentation
I've finally changed the nightmare that this module was (with everything
being inside of the main file), now it's a nightmare divided between
several modules.

I've tried to turn the workers and the thread pool into a
more generic implementation, but since they require an async function to
work with, this turned out to be almost impossible. While trying to
implement it I've ended up with a pinned box with a dynamic Future with
the output I had, and I just couldn't handle seeing any more of the `>`
arrows and trying to make it all just work.

Yeah, now it all can work only together with each other, and I will
probably have to rework it a little once we start working with more
complex and different scrappers, but at least it works for now.

Also updated a couple of TODOs, and added some documentation.

So, with three tasks done, now only three top-priority things remain:
* Add tests
* Test the module itself, figure out how many threads are useful, and
how many webpages should I send to each thread before it becomes too
overwhelmed
* Find a nice way to provide progress on each thread in the command line
